functionality for reporting temp (converting from C to F), adding in src/header for StepCounter class

+++ microbit-v2-samples/source/health_tracker/include/HealthSensor.h
@@ -8,7 +8,7 @@ public:
     virtual ~HealthSensor() {}  // Virtual destructor (important for polymorphism)
     
     // Pure virtual function to be overridden by child classes
-    virtual float readData() = 0;  
+    virtual int readData() = 0;  
 };
 
 #endif // HEALTHSENSOR_H
